+++ pkg/cmd/task/edit/edit_test.go
@@ -94,9 +94,7 @@ func TestCmdEdit(t *testing.T) {
 				f.On("GetWorkspaceID").
 					Return("w", nil)
 
-				cf := mocks.NewMockConfig(t)
-				f.On("Config").Return(cf)
-				cf.On("IsAllowNameForID").Return(false)
+				f.EXPECT().Config().Return(&mocks.SimpleConfig{})
 
 				f.On("Client").Return(nil, errors.New("client error"))
 				return f
@@ -147,9 +145,9 @@ func TestCmdEdit(t *testing.T) {
 					Return("w", nil)
 				f.On("Client").Return(c, nil)
 
-				cf := mocks.NewMockConfig(t)
-				f.On("Config").Return(cf)
-				cf.On("IsAllowNameForID").Return(true)
+				f.EXPECT().Config().Return(&mocks.SimpleConfig{
+					AllowNameForID: true,
+				})
 
 				c.On("GetProjects", api.GetProjectsParam{
 					Workspace:       "w",
@@ -177,9 +175,9 @@ func TestCmdEdit(t *testing.T) {
 					Return("w", nil)
 				f.On("Client").Return(c, nil)
 
-				cf := mocks.NewMockConfig(t)
-				f.On("Config").Return(cf)
-				cf.On("IsAllowNameForID").Return(true)
+				f.EXPECT().Config().Return(&mocks.SimpleConfig{
+					AllowNameForID: true,
+				})
 
 				c.On("GetProjects", api.GetProjectsParam{
 					Workspace:       "w",
@@ -239,9 +237,9 @@ func TestCmdEdit(t *testing.T) {
 					Return("w", nil)
 				f.On("Client").Return(c, nil)
 
-				cf := mocks.NewMockConfig(t)
-				f.On("Config").Return(cf)
-				cf.On("IsAllowNameForID").Return(true)
+				f.EXPECT().Config().Return(&mocks.SimpleConfig{
+					AllowNameForID: true,
+				})
 
 				c.On("GetProjects", api.GetProjectsParam{
 					Workspace:       "w",
@@ -290,9 +288,9 @@ func TestCmdEdit(t *testing.T) {
 					Return("w", nil)
 				f.On("Client").Return(c, nil)
 
-				cf := mocks.NewMockConfig(t)
-				f.On("Config").Return(cf)
-				cf.On("IsAllowNameForID").Return(true)
+				f.EXPECT().Config().Return(&mocks.SimpleConfig{
+					AllowNameForID: true,
+				})
 
 				c.On("GetProjects", api.GetProjectsParam{
 					Workspace:       "w",