+++ pkg/cmd/time-entry/report/util/reportwithrange_test.go
@@ -117,6 +117,7 @@ func TestReportWithRange(t *testing.T) {
 				cf := mocks.NewMockConfig(t)
 				f.On("Config").Return(cf)
 				cf.On("IsAllowNameForID").Return(true)
+				cf.On("IsSearchProjectWithClientsName").Return(false)
 
 				c := mocks.NewMockClient(t)
 				f.On("Client").Return(c, nil)
@@ -174,6 +175,7 @@ func TestReportWithRange(t *testing.T) {
 				cf := mocks.NewMockConfig(t)
 				f.On("Config").Return(cf)
 				cf.On("IsAllowNameForID").Return(true)
+				cf.On("IsSearchProjectWithClientsName").Return(false)
 
 				c := mocks.NewMockClient(t)
 				f.On("Client").Return(c, nil)
@@ -210,6 +212,7 @@ func TestReportWithRange(t *testing.T) {
 				cf := mocks.NewMockConfig(t)
 				f.On("Config").Return(cf)
 				cf.On("IsAllowNameForID").Return(true)
+				cf.On("IsSearchProjectWithClientsName").Return(false)
 
 				c := mocks.NewMockClient(t)
 				f.On("Client").Return(c, nil)